+++ framework/hyper-express/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @forklaunch/hyper-express
 
+## 1.2.25
+
+### Patch Changes
+
+- sync changes across packages
+- Updated dependencies
+  - @forklaunch/validator@1.2.12
+  - @forklaunch/common@1.2.12
+  - @forklaunch/core@1.3.15
+  - @forklaunch/ws@1.2.25
+
+## 1.2.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @forklaunch/core@1.3.14
+  - @forklaunch/ws@1.2.24
+
 ## 1.2.23
 
 ### Patch Changes